+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added the date range component to the holdings tab
 
+### Changed
+
+- Extended the statistics section on the about page (users in Slack community)
+
 ### Fixed
 
 - Fixed the creation of historical data in the admin control panel (upsert instead of update)